refactor audiocaps
